Add expand collections optimization
- Add getExpandedCollection method for efficient API calls
- Support expanded collections for all resource types
- Add collection structs for expanded responses
- Fix GPU metric type casting issues
- Improve network adapter port handling

+++ internal/collector/metrics.go
@@ -782,7 +782,7 @@ func (mc *Collector) NewGpuPowerConsumedWatts(ch chan<- prometheus.Metric, m *GP
 	ch <- prometheus.MustNewConstMetric(
 		mc.GpuPowerConsumedWatts,
 		prometheus.GaugeValue,
-		float64(m.PowerConsumedWatts),
+		float64(int32(m.PowerConsumedWatts)),
 		m.Id,
 	)
 }
@@ -791,7 +791,7 @@ func (mc *Collector) NewGpuTemp(ch chan<- prometheus.Metric, m *GPU) {
 	ch <- prometheus.MustNewConstMetric(
 		mc.GpuTemp,
 		prometheus.GaugeValue,
-		float64(m.TemperatureCelsius),
+		float64(int32(m.TemperatureCelsius)),
 		m.Id,
 	)
 }
